Add files via upload
This is the original website which has been largely influenced by the work of Jonas Schmedtmann. I followed his tutorial's on Udemy to create this website and adapted it to suit my own needs with the help of W3Schools and various YouTube tutorials.

Much thanks to these great people!